Dtype changes

+++ src/models/intermediate_robust_generator/model.py
@@ -28,7 +28,7 @@ class generator_model(tfk.Model):
 
     def _max_proba(self, proba):
         totals = tfm.reduce_sum(proba, axis=0)
-        probs = tf.map_fn(lambda x : tf.one_hot(tf.argmax(x), depth=x.shape[0], on_value=1, off_value=0), elems=totals)
+        probs = tf.map_fn(lambda x : tf.one_hot(tf.argmax(x), depth=x.shape[0], on_value=1, off_value=0, dtype=tf.float32), elems=totals)
 
         return probs
